refactor: move picked by fetch to subgraph client
