+++ src/clients/SnapshotSubgraph.ts
@@ -1,6 +1,7 @@
 import fetch from 'isomorphic-fetch'
 
 import { SNAPSHOT_QUERY_ENDPOINT } from '../entities/Snapshot/constants'
+import { PICKED_BY_QUERY } from '../entities/Snapshot/queries'
 
 import { Delegation } from './SnapshotGraphql'
 import { trimLastForwardSlash } from './utils'
@@ -8,6 +9,11 @@ import { trimLastForwardSlash } from './utils'
 export type DelegationQueryResult = {
   delegatedTo: Delegation[]
   delegatedFrom: Delegation[]
+}
+
+export type PickedByResult = {
+  address: string
+  pickedBy: number
 }
 
 export class SnapshotSubgraph {
@@ -72,5 +78,62 @@ export class SnapshotSubgraph {
     }
 
     return delegations
+  }
+
+  async getPickedBy(variables: { address: string[]; space: string }) {
+    const { address, space } = variables
+
+    if (!address || !space) {
+      return []
+    }
+
+    let hasNext = true
+    let skip = 0
+    const first = 500
+
+    let delegations: Delegation[] = []
+    while (hasNext) {
+      const response = await fetch(SNAPSHOT_QUERY_ENDPOINT, {
+        method: 'post',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({
+          query: PICKED_BY_QUERY,
+          variables: { ...variables, skip, first },
+        }),
+      })
+
+      const body = await response.json()
+      const currentDelegations = body?.data?.delegatedFrom || []
+      delegations = [...delegations, ...currentDelegations]
+
+      if (currentDelegations.length < first) {
+        hasNext = false
+      } else {
+        skip = delegations.length
+      }
+    }
+
+    const pickedBy = new Map<string, Set<string>>()
+
+    for (const addr of address) {
+      const filteredDelegations = delegations.filter((deleg) => deleg.delegate === addr)
+      pickedBy.set(addr, new Set())
+
+      if (filteredDelegations.length > 0) {
+        for (const deleg of filteredDelegations) {
+          pickedBy.get(deleg.delegate)?.add(deleg.delegator)
+        }
+      }
+    }
+
+    const result: PickedByResult[] = []
+
+    for (const entry of pickedBy.entries()) {
+      const address = entry[0]
+      const pickedBy = entry[1].size
+      result.push({ address, pickedBy })
+    }
+
+    return result
   }
 }
